requirements.txt
coreapi==2.3.3
Django==3.1.7
djangorestframework==3.12.2
